% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_nowcast2.R
\name{aux_nowcast2}
\alias{aux_nowcast2}
\title{Auxiliar function for nowcast}
\usage{
aux_nowcast2(y, x)
}
\arguments{
\item{y}{Dependent variable.}

\item{x}{Independent variable.}
}
\description{
Forecast quarterly series from monthly series.
}
