% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRGDP.R
\docType{data}
\name{BRGDP}
\alias{BRGDP}
\title{Real Time Database of 100 economic variables from 12/11/2018}
\format{The format is:
List of 4
$ base : Time-Series [1:227, 1:100] from 2000 to 2019: NA NA NA NA NA NA NA NA NA NA ...
..- attr(*, "dimnames")=List of 2
.. ..$ : NULL
.. ..$ : chr [1:100] "DIVIDA_6" "IBC_BR" "PMC_TOTAL" "PIM_BCONSND" ...
$ trans: num [1:100] 5 3 3 3 3 3 3 3 3 3 ...
$ delay: num [1:100] 14 42 42 42 42 42 35 35 35 35 ...
$ GDP  : Time-Series [1:75] from 2000 to 2018: 106 110 114 114 110 ...}
\usage{
BRGDP
}
\description{
a list containing the database, a vector of transformation for the function Bpanel, a vector of delays for the function PRTDB, and the GDP time series.
}
\keyword{datasets}
