% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{bridge}
\alias{bridge}
\title{Bridge Equation}
\usage{
bridge(y, x)
}
\arguments{
\item{y}{Dependent variable.}

\item{x}{Independent variable.}
}
\description{
Bridge equation, estimate quarterly series from monthly series.
}
