% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorExtraction.R
\name{FactorExtraction}
\alias{FactorExtraction}
\title{Factor Extraction}
\usage{
FactorExtraction(x = NULL, q = NULL, r = NULL, p = NULL, A = NULL,
  C = NULL, Q = NULL, R = NULL, initx = NULL, initV = NULL,
  ss = NULL, MM = NULL, n.prevs = NULL)
}
\arguments{
\item{x.}{none}

\item{q.}{none}

\item{r.}{none}

\item{p.}{none}

\item{A.}{none}

\item{C.}{none}

\item{Q.}{none}

\item{R.}{none}

\item{initx.}{none}

\item{initV.}{none}

\item{ss.}{none}

\item{MM.}{none}

\item{n.prevs.}{none}
}
\description{
Extract factors from a data set of time series.
}
