% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecasting}
\alias{forecasting}
\title{Forecasting}
\usage{
forecasting(y, x)
}
\arguments{
\item{y.}{Variável de interesse, trimestre}

\item{x.}{Variáveis mensais usadas para fazer a previsão}
}
\description{
Faz previsão de variável trimestral a partir de variáveis mensais. Aqui utilizamos os fatores como variáveis mensais.
}
