% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_factors.R
\name{IC_factors}
\alias{IC_factors}
\title{Information criterium for number of factors}
\usage{
IC_factors(x, rmax = NULL, type = 2)
}
\arguments{
\item{x}{a dataset;}

\item{rmax}{a positive integer corresponding to the maximum number of factors for which the information criterium should be tested;}

\item{type}{a positive integer corresponding to the chosen information criterium (1,2,3). The default is 2.}
}
\value{
A \code{list} containing two elements:

\item{r_star}{The number of factors minimizing the information criterium;}
\item{IC}{A vector of values of the information criterium for all the tested number of factors.}
}
\description{
Minimizes the selected information criterium to determine the number of factors to be used in an approximate factor model.
}
\references{
Bai, J., Ng, S. (2002). Determining the Number of Factors in Approximate Factor Models. Econometrica, 70(1), 191-221. <doi:10.1111/1468-0262.00273>
}
