% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_smoother_diag.R
\name{kalman_smoother_diag}
\alias{kalman_smoother_diag}
\title{Kalman Smoother Diagonal}
\usage{
kalman_smoother_diag(y, A, C, Q, R, init_x, init_V, varagin)
}
\arguments{
\item{y.}{none}

\item{A.}{none}

\item{C.}{none}

\item{Q.}{none}

\item{R.}{none}

\item{init_x.}{none}

\item{init_V.}{none}

\item{varagin.}{none}
}
\description{
Kalman Smoother Diagonal
}
