% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_update_diag.R
\name{kalman_update_diag}
\alias{kalman_update_diag}
\title{Kalman update diagonal}
\usage{
kalman_update_diag(A, C, Q, R, y, x, V, varagin)
}
\arguments{
\item{A.}{none}

\item{C.}{none}

\item{Q.}{none}

\item{R.}{none}

\item{y.}{none}

\item{x.}{none}

\item{V.}{none}

\item{varagin.}{none}
}
\description{
Kalman update diagonal
}
