% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month2qtr.R
\name{month2qtr}
\alias{month2qtr}
\title{Monthly to quarterly agregation}
\usage{
month2qtr(x)
}
\arguments{
\item{x}{Variable in month frequency}
}
\value{
The correpondent quarterly transformation or agregation.
}
\description{
It transforsms a monthly time series in a quarterly, selecting the last month to represent the quarter. So its necessary to agregate the original serie before use this function.
}
\examples{
# Selecting only last month of matrix time series BRGDP:
mestri_vintage<-month2qtr(BRGDP)

# Selecting only last month of first two columns of matrix time series BRGDP:
mestri_vintage<-month2qtr(BRGDP[,1:2])

}
