% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.plot.R
\name{nowcast.plot}
\alias{nowcast.plot}
\title{Plot for nowcast output function}
\usage{
nowcast.plot(out, type = "fcst")
}
\arguments{
\item{out}{Output of function nowcast}

\item{type}{Type of the function}
}
\description{
Make plot to visualize the output of nowcast function
}
\examples{
\dontrun{
trans <- USGDP$Legenda$Transformation[-length(USGDP$Legenda$Transformation)]
base <- USGDP$Base[,-dim(USGDP$Base)[2]]
gdp <- month2qtr(USGDP$Base[,dim(USGDP$Base)[2]])
base <- Bpanel(base = base, trans = trans)
now <- nowcast(y = gdp, x = base, q = 2, r = 3)

nowcast.plot(now, type = "fcst")
nowcast.plot(now, type = "factors")
nowcast.plot(now, type = "eigenvalues")
nowcast.plot(now, type = "eigenvectors")
}
}
