% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting.R
\name{nowcasting}
\alias{nowcasting}
\title{Nowcasting of a quarterly time serie using a dynamic factor.}
\usage{
nowcasting(y, vintage, legenda, q = 2, r = 2, p = 1)
}
\arguments{
\item{y.}{Quarterly time-series}

\item{base.}{A time series matrix (mts) representing the vintage of interest.}

\item{legenda.}{\code{data.frame} or \code{vector}. A \code{data.frame} with two columns, the first one is the name, and the second is the transformation to let the series become stationary.
A \code{vector} where each coordinate is the transformation of the correspondent coordinate in the \code{mts} of the previous argument. 
The transformation is specified as follow:
# trans 0, the original serie is preserved.
# trans 1 \deqn{latex1}{100*\frac{X_t - X_{t-1}}{X_{t-1}}} 
# trans 2 \deqn{latex2}{X_t - X_{t-1}}
# trans 3 \deqn{latex3}{100*\frac{X_t - X_{t-12}}{X_{t-12}}  -  100*\frac{X_{t-1} - X_{t-13}}{X_{t-13}}}}

\item{q.}{Dynamic rank. Number of error terms. If not specified q = 2.}

\item{r.}{Static rank (r>=q), i.e. number of factors. If not specified r = 2.}

\item{p.}{AR order of factors. If not specified p = 1.}
}
\description{
Estimate nowcasting and foreacasting for a quarterly time serie. The method is based on 
\emph{Giannone, Domenico, Lucrezia Reichlin, and David Small. "Nowcasting: The real-time informational content of macroeconomic data." Journal of Monetary Economics 55.4 (2008): 665-676.}
}
\seealso{
\code{\link[nowcasting]{base_extraction}}
}
