% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_correction.R
\name{outliers_correction}
\alias{outliers_correction}
\title{Outliers Correction}
\usage{
outliers_correction(x, k_ma = 3)
}
\arguments{
\item{x.}{none}

\item{k_ma.}{none}
}
\description{
It corrects outliers
}
