% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatodfm.R
\name{pcatodfm}
\alias{pcatodfm}
\title{Principal Components to Dynamic Factor Model}
\usage{
pcatodfm(x, q, r, p)
}
\arguments{
\item{x}{The \code{ts} used to extract the dynamic factors.}

\item{q}{Number of shocks in factors.}

\item{r}{Number of factors.}

\item{p}{Degree of autoregressive polynomial}
}
\description{
Estimates the first stage for dynamic factors on \emph{Giannone, Domenico, Lucrezia Reichlin, and David Small. "Nowcasting: The real-time informational content of macroeconomic data." Journal of Monetary Economics 55.4 (2008): 665-676}.
}
