% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTDB.R
\name{RTDB}
\alias{RTDB}
\title{Re-create Real Time Data Base}
\usage{
RTDB(series_code = NULL, vintage = NULL)
}
\arguments{
\item{series_code}{Vector with the series encoding follow the Bacen (Banco Central do Brasil) standards.}

\item{vintage}{The vintage encoded by the day of the extraction}
}
\description{
Create a time series matrix \code{mts} replicating the information available in a given date.
}
\examples{
# Extracting GDP serie at real-time from Central Bank of Brasil data base
gdp<-base_extraction(22099)
}
\references{
Central Bank of Brazil
}
