% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USGDPshort.R
\docType{data}
\name{USGDPshort}
\alias{USGDPshort}
\title{Example of replication files in Banbura et al. 2011}
\format{A \code{list} with 2 elements: 

\itemize{
\item \code{base} is a \code{mts} with 26 series and 358 observations. There are missing values;
\item \code{legend} is a \code{data.frame} with specifications of the series in USGDPshort$base.
}}
\source{
This dataset is available as \emph{replication files} of the seminal work \emph{Banbura et al 2011}.
 One can find these \emph{replication files} in the following url: \url{https://www.newyorkfed.org/research/economists/giannone/pub}
}
\usage{
USGDPshort
}
\description{
Dataset available to replicate the results in \emph{Banbura et al. 2011}.
}
\references{
Banbura, M., Giannone, D. & Reichlin, L. (2011). Nowcasting. Oxford Handbook on Economic Forecasting, ed. by M. P. Clements, and D. F. Hendry, pp. 63-90. Oxford University Press.
}
