% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrumarVintage.R
\name{arrumarVintage}
\alias{arrumarVintage}
\title{Balanced pannel}
\usage{
arrumarVintage(base = NULL, legenda = NULL)
}
\arguments{
\item{base.}{The vintage}

\item{legenda.}{The  legends}
}
\description{
Create a balanced pannel for data. Select variables, and transform time series to its stationary representation
}
