% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_extraction.R
\name{base_extraction}
\alias{base_extraction}
\title{Extract time series and create a base}
\usage{
base_extraction(series_code)
}
\arguments{
\item{series_code.}{Vector with the series encoding follow the Bacen (Banco Central do Brasil) standards.}
}
\value{
A \code{mts} in the same specification required in argument \code{base} of function \code{nowcasting}
}
\description{
Create a time series matrix \code{mts} used to estimate the commmon factors.
}
\seealso{
\code{\link[BETS]{BETS.get}}
}
