% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRGDP.R
\docType{data}
\name{BRGDP}
\alias{BRGDP}
\title{Real Time Data Base Extracted in 03/10/2017}
\format{The format is:
Time-Series [1:278, 1:7] from 1994 to 2018: 0.933 0.899 0.865 0.846 0.842 ...
- attr(*, "dimnames")=List of 2
..$ : NULL
..$ : chr [1:7] "serie1" "serie12" "serie1373" "serie1453" ...}
\usage{
BRGDP
}
\description{
This is a panel containning a piece of the information set of a specific date for Brazilian economic activity. The panel is formated as a matrix time series \code{mts} containing: Exchange rate - Free - United States dollar (1); Interest rate - CDI (12); Vehicles production (1373); Credit Sales Index (1453); Retail sales (1455); Current economic conditions index (4394); Industrial production, general index (21859); Quarterly GDP - observed data - GDP at market prices (22099).
}
\keyword{datasets}
