% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecasting}
\alias{forecasting}
\title{Forecasting}
\usage{
forecasting(y, x)
}
\arguments{
\item{y}{Dependent variable.}

\item{x}{Independent variable.}
}
\description{
Forecast quarterly series from monthly series.
}
