% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_diag.R
\name{kalman_filter_diag}
\alias{kalman_filter_diag}
\title{Kalman Filter Diagonal}
\usage{
kalman_filter_diag(y, A, C, Q, R, init_x, init_V, varagin)
}
\arguments{
\item{y}{xxx}

\item{A}{Matrix that update factors with VAR}

\item{C}{Matrix that combine factors to explain the transformed data.}

\item{Q}{Error variance in factor update.}

\item{R}{Error variance in explain data from factors}

\item{init_x}{xxx}

\item{init_V}{xxx}

\item{varagin}{xxx}
}
\description{
Kalman Filter Diagonal. Function translated from matlab and transformed from Giannone et al. (2008) replication files.
}
