% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_update_diag.R
\name{kalman_update_diag}
\alias{kalman_update_diag}
\title{Kalman update diagonal}
\usage{
kalman_update_diag(A, C, Q, R, y, x, V, varagin)
}
\arguments{
\item{A}{Matrix that update factors with VAR}

\item{C}{Matrix that combine factors to explain the transformed data.}

\item{Q}{Error variance in factor update.}

\item{R}{Error variance in explain data from factors}

\item{y}{xxx}

\item{x}{xxx}

\item{V}{xxx}

\item{varagin}{xxx}
}
\description{
Kalman update diagonal. This function is a translation and modification of a function with the same name, written in MATLAB, and available in the replication files of the Giannone et al. (2008) <doi: 10.1016 / j.jmoneco.2008.05.010>. One can find this files at \url{https://www.newyorkfed.org/research/economists/giannone/pub}.
}
