% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_update_diag.R
\name{kalman_update_diag}
\alias{kalman_update_diag}
\title{Kalman update diagonal}
\usage{
kalman_update_diag(A, C, Q, R, y, x, V, varagin)
}
\arguments{
\item{A}{Matrix that update factors with VAR}

\item{C}{Matrix that combine factors to explain the transformed data.}

\item{Q}{Error variance in factor update.}

\item{R}{Error variance in explain data from factors}

\item{y}{xxx}

\item{x}{xxx}

\item{V}{xxx}

\item{varagin}{xxx}
}
\description{
Kalman update diagonal
}
