% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mestri.R
\name{mestri}
\alias{mestri}
\title{Month to quarter agregation}
\usage{
mestri(x)
}
\arguments{
\item{x.}{Variable in month frequency}
}
\value{
The correpondent quarterly transformation or agregation.
}
\description{
It transforsms a monthly time series in a quarterly.
}
