% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monqua.R
\name{monqua}
\alias{monqua}
\title{Monthly to quarterly agregation}
\usage{
monqua(x)
}
\arguments{
\item{x}{Variable in month frequency}
}
\value{
The correpondent quarterly transformation or agregation.
}
\description{
It transforsms a monthly time series in a quarterly, selecting the last month to represent the quarter. So its necessary to agregate the original serie before use this function.
}
\examples{
# Selecting only last month of matrix time series vintage:
mestri_vintage<-monqua(vintage)

# Selecting only last month of first two columns of matrix time series vintage:
mestri_vintage<-monqua(vintage[,1:2])

}
