% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month2qtr.R
\name{month2qtr}
\alias{month2qtr}
\title{Monthly to quarterly transformation}
\usage{
month2qtr(x, reference_month = 3)
}
\arguments{
\item{x}{a \code{ts} or \code{mts} in monthly frequency}

\item{reference_month}{an integer to define the reference month that will represent the quarter. Default is 3. The options are 1, 2 or 3.}
}
\value{
The correspondent quarterly transformation.
}
\description{
It transforms a monthly time series in a quarterly one, selecting the last month of the quarter to represent the value of the quarter.
}
\examples{
# Selecting only last month of matrix time series BRGDP:
month2qtr(BRGDP)

# Vehicle production in the quarter from vehicle production in the month
month2qtr(stats::filter(BRGDP[,3],c(1,1,1),sides=1))

}
