% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcasting of a quarterly time serie using a dynamic factor.}
\usage{
nowcast(y, x, q = 2, r = 2, p = 1, method = "2sq", blocks = NULL)
}
\arguments{
\item{y}{Stationary quarterly time-series}

\item{x}{A time series matrix (\code{mts}) representing the regressors of interest. The series must be stationary.}

\item{q}{Dynamic rank. Number of error terms. If not specified q = 2.}

\item{r}{Static rank (r>=q), i.e. number of factors. If not specified r = 2.}

\item{p}{AR order of factors. If not specified p = 1.}

\item{method}{2sq: Two stages quarterly for Giannone et al. 2008; 2sm: Two stages monthly for Banbura and Runstler 2011; EM: Expected Maximization Giannone et al 2011}

\item{blocks}{only for EM method. Select which factors impact the variables (global, nominal or real).}
}
\value{
A \code{list} containing two elements:

A \code{mts} named \code{main} contains the original serie, the estimation in the sample, the estimation out of the sample;

A \code{list} named \code{factors} contains the estimated factors and coeffients.

A \code{mts} named \code{fore_x} contains the output of all regressors.

A \code{mts} named \code{monthgdp} contains the a monthly measure for GDP.
}
\description{
Estimate nowcasting and foreacasting for a quarterly time serie.
}
\examples{
\dontrun{
pib<-base_extraction(22099)
retpib2<-month2qtr(diff(diff(lag(pib,-2),3),12))
base1<-Bpanel(BRGDP,rep(4,dim(BRGDP)[2]),aggregate = T)
now1<-nowcast(retpib2,base1,1,1,1,method = '2sq')

base2<-Bpanel(BRGDP,rep(4,dim(BRGDP)[2]),aggregate = F)
now2<-nowcast(retpib2,base2,2,2,1,'2sm')

now3<-nowcast(retpib2,base2,1,1,1,'EM')

}
}
\references{
Giannone, D., Reichlin, L., & Small, D. (2008). Nowcasting: The real-time informational content of macroeconomic data. Journal of Monetary Economics, 55(4), 665-676.<doi:10.1016/j.jmoneco.2008.05.010>
}
\seealso{
\code{\link[nowcasting]{base_extraction}}
}
