% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcasting of a quarterly time serie using a dynamic factor.}
\usage{
nowcast(y, x, q = 2, r = 2, p = 1, method = "2sq", blocks = NULL)
}
\arguments{
\item{y}{Stationary quarterly time-series}

\item{x}{A time series matrix (\code{mts}) representing the regressors of interest. The series must be stationary.}

\item{q}{Dynamic rank. Number of error terms. If not specified q = 2.}

\item{r}{Static rank (r>=q), i.e. number of factors. If not specified r = 2.}

\item{p}{AR order of factors. If not specified p = 1.}

\item{method}{2sq: Two stages quarterly for Giannone et al. 2008; 2sm: Two stages monthly for Banbura and Runstler 2011; EM: Expected Maximization Giannone et al 2011}

\item{blocks}{only for EM method. Select which factors impact the variables (global, nominal or real).}
}
\value{
A \code{list} containing two elements:

A \code{mts} named \code{main} contains the original serie, the estimation in the sample, the estimation out of the sample;

A \code{list} named \code{factors} contains the estimated factors and coeffients.

A \code{mts} named \code{fore_x} contains the output of all regressors.

A \code{mts} named \code{monthgdp} contains the a monthly measure for GDP.
}
\description{
Estimate nowcasting and foreacasting for a quarterly time serie.
}
\examples{
\dontrun{
# nowcast function examples:

### Method 2sq
### This is two stages quarterly, based on Giannone et al. (2008)
pib<-BRGDP[,8]
y<-month2qtr(diff(diff(pib,3),12))
x<-Bpanel(BRGDP[,-8],rep(4,dim(BRGDP)[2]),aggregate = T)
q<-1
r<-1
p<-1
now_2sq<-nowcast(y,x,q,r,p,method = '2sq')
# outputs
nowcast.plot(now_2sq)
summary(now_2sq$reg)
nowcast.plot(now_2sq,type = 'factors')


### Method 2sm
### This is two stages monthly, based on Giannone et al. (2008) and Banbura and Runstler (2011)
pib<-BRGDP[,8]
y<-month2qtr(diff(diff(pib,3),12))
x<-Bpanel(BRGDP[,-8],rep(4,dim(BRGDP)[2]),aggregate = F)
q<-1
r<-2
p<-1
now_2sm<-nowcast(y,x,q,r,p,method = '2sm')
# outputs
nowcast.plot(now_2sm)
summary(now_2sm$reg)
nowcast.plot(now_2sm,type = 'factors')
ts.plot(now_2sm$monthgdp)


### Method EM
y<-month2qtr(diff(diff(pib,3),12))
x<-Bpanel(BRGDP[,-8],rep(4,dim(BRGDP)[2]),aggregate = F)
q=1
r=2
p=1
now_em<-nowcast(y,x,q,r,p,'EM')
# outputs
nowcast.plot(now_em)
ts.plot(now_em$monthgdp)
}
}
\references{
Giannone, D., Reichlin, L., & Small, D. (2008). Nowcasting: The real-time informational content of macroeconomic data. Journal of Monetary Economics, 55(4), 665-676.<doi:10.1016/j.jmoneco.2008.05.010>
}
\seealso{
\code{\link[nowcasting]{base_extraction}}
}
