% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting.R
\name{nowcasting}
\alias{nowcasting}
\title{Nowcasting of a quarterly time serie using a dynamic factor.}
\usage{
nowcasting(y, vintage, legenda, q = 2, r = 2, p = 1)
}
\arguments{
\item{y.}{Quarterly time-series}

\item{vintage.}{A time series matrix \(mts\) representing the vintage of interest.}

\item{legenda.}{A data.frame with two columns, the first one is the name, and the second is the transformation to let the series become stationary.}

\item{q.}{Dynamic rank. Number of error terms. If not specified q = 2.}

\item{r.}{Static rank \(r>=q\), i.e. number of factors. If not specified r = 2.}

\item{p.}{AR order of factors. If not specified p = 1.}
}
\description{
This is a R translation of MATLAB code presented in the following paper: Giannone, Domenico, Lucrezia Reichlin, and David Small. "Nowcasting: The real-time informational content of macroeconomic data." Journal of Monetary Economics 55.4 (2008): 665-676.
}
