% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting.R
\docType{package}
\name{nowcasting}
\alias{nowcasting}
\alias{nowcasting-package}
\title{Nowcasting Analysis and Create Real-Time Data Basis}
\description{
teste
}
\note{
The authors would like to thank the support by the Getulio 
Vargas Foundation (FGV).
}
\author{
Daiane Mattos \email{daiane.mattos@fgv.br},
        Guilherme Branco \email{guilherme.branco@fgv.br},
        Pedro Costa Ferreira \email{pedro.guilherme@fgv.br}
}
