% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_correction.R
\name{outliers_correction}
\alias{outliers_correction}
\title{Outliers Correction}
\usage{
outliers_correction(x, k_ma = 3)
}
\arguments{
\item{x}{A \code{ts} to be adjusted.}

\item{k_ma}{Moving average order.}
}
\description{
This function operates in two stages. First, it substitutes missings and outliers for the median
Outliers are definited as observations that its distance from median is more than 4 times the IQR (Interquartile Range).
Second, a \code{k_ma} moving average is filtered in each serie, and its values are used to replace outliers and missings in the original data set.
}
