% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_correction.R
\name{outliers_correction}
\alias{outliers_correction}
\title{Outliers Correction}
\usage{
outliers_correction(x, k_ma = 3)
}
\arguments{
\item{x.}{Um objeto \code{ts} que se deseja aplicar a transformação de substituição de outliers e missings.}

\item{k_ma.}{Ordem da média móvel utilizada para substituição das variávesi}
}
\description{
Essa função opera em dois estágios. Inicialmente substitui os missings e outliers pela mediana da série.
Os outliers são definidos como as observações que distão da mediana mais de 4 vezes o IQR. No segundo estágio é feita uma média móvel de ordem \code{k_ma}, os valores iniciais de missings e outliers são substituídos pelo resultado dessa média móvel.
}
