% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatodfm.R
\name{pcatodfm}
\alias{pcatodfm}
\title{Principal Components to Dynamic Factor Model}
\usage{
pcatodfm(x, q, r, p)
}
\arguments{
\item{x.}{base de dados em formato ts}

\item{q.}{o número de choques nos fatores}

\item{r.}{a quantidade de fatores}

\item{p.}{o grau do polinômio autorregressivo}
}
\description{
This function estimates the first stage in Giannone et al. \(2008\) model.
}
