% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatodfm.R
\name{pcatodfm}
\alias{pcatodfm}
\title{Principal Components to Dynamic Factor Model}
\usage{
pcatodfm(x, q, r, p)
}
\arguments{
\item{x.}{none}

\item{q.}{none}

\item{r.}{none}

\item{p.}{none}
}
\description{
This function estimates the first stage in Giannone et al. \(2008\) model.
}
