% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr2month.r
\name{qtr2month}
\alias{qtr2month}
\title{Quarterly to monthly transformation}
\usage{
qtr2month(x)
}
\arguments{
\item{x}{a \code{ts} or \code{mts} in quaterly frequency}
}
\value{
The correspondent monthly transformation.
}
\description{
It transforms a quarterly time series in a monthly one. To the last month of the quarter is set the value of the quarter.
}
\examples{
# Selecting only last month of matrix time series BRGDP:
mestri_vintage<-month2qtr(BRGDP[,dim(BRGDP)[2]])

qtr2month(mestri_vintage) 

}
