% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr2month.r
\name{qtr2month}
\alias{qtr2month}
\title{Quarterly to monthly desagregation}
\usage{
qtr2month(x)
}
\arguments{
\item{x}{Variable in quarterly frequency}
}
\value{
The correpondent monthly transformation or agregation.
}
\description{
It transforsms a quarterly time series in a monthly. To the last month of the monthly \code{ts} is set the value of the quarterly \code{ts}.
}
\examples{
# Selecting only last month of matrix time series BRGDP:
mestri_vintage<-month2qtr(BRGDP[,dim(BRGDP)[2]])

qtr2month(mestri_vintage) 

}
