% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr2month.r
\name{qtr2month}
\alias{qtr2month}
\title{Quarterly to monthly transformation}
\usage{
qtr2month(x)
}
\arguments{
\item{x}{a \code{ts} or \code{mts} in quarterly frequency}
}
\value{
The correpondent monthly transformation.
}
\description{
It transforms quarterly a  time series in a monthly one. The values of the quarterly \code{ts} are set to the last month of the quarter.
}
\examples{
# Selecting the quarterly GDP variable in BRGDP
brgdp <- month2qtr(BRGDP[,ncol(BRGDP)])

qtr2month(brgdp) 

}
