% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_update.R
\name{smooth_update}
\alias{smooth_update}
\title{Smooth update}
\usage{
smooth_update(xsmooth_future, Vsmooth_future, xfilt, Vfilt, Vfilt_future,
  VVfilt_future, A, Q, B, u)
}
\arguments{
\item{xsmooth_future}{xxx}

\item{Vsmooth_future}{xxx}

\item{xfilt}{xxx}

\item{Vfilt}{xxx}

\item{Vfilt_future}{xxx}

\item{VVfilt_future}{xxx}

\item{A}{Matrix that update factors with VAR}

\item{Q}{Error variance in factor update.}

\item{B}{xxx}

\item{u}{xxx}
}
\description{
Smooth update. This function is a translation and modification of a function with the same name, written in MATLAB, and available in the replication files of the Giannone et al. (2008) <doi: 10.1016 / j.jmoneco.2008.05.010>. One can find this files at \url{https://www.newyorkfed.org/research/economists/giannone/pub}.
}
