% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_update.R
\name{smooth_update}
\alias{smooth_update}
\title{Smooth update}
\usage{
smooth_update(xsmooth_future, Vsmooth_future, xfilt, Vfilt, Vfilt_future,
  VVfilt_future, A, Q, B, u)
}
\arguments{
\item{xsmooth_future.}{none}

\item{Vsmooth_future.}{none}

\item{xfilt.}{none}

\item{Vfilt.}{none}

\item{Vfilt_future.}{none}

\item{VVfilt_future.}{none}

\item{A.}{none}

\item{Q.}{none}

\item{B.}{none}

\item{u.}{none}
}
\description{
Smoothe update
}
